+++ stars-logic-kcmftbl/src/main/kotlin/tools/aqua/stars/logic/kcmftbl/firstorder/bind.kt
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package tools.aqua.stars.logic.kcmftbl
+package tools.aqua.stars.logic.kcmftbl.firstorder
 
 /**
  * CMFTBL implementation of the 'bind' operator.