+++ stars-logic-kcmftbl/src/test/kotlin/tools/aqua/stars/logic/kcmftbl/formulas/TestGlobally.kt
@@ -21,6 +21,8 @@ import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
 import tools.aqua.stars.logic.kcmftbl.*
+import tools.aqua.stars.logic.kcmftbl.future.eventually
+import tools.aqua.stars.logic.kcmftbl.future.globally
 
 /** This class tests the CMFTBL operator [globally]. */
 class TestGlobally {