+++ stars-logic-kcmftbl/src/test/kotlin/tools/aqua/stars/logic/kcmftbl/formulas/TestMaxPrevalence.kt
@@ -22,6 +22,7 @@ import kotlin.test.assertFailsWith
 import kotlin.test.assertFalse
 import kotlin.test.assertTrue
 import tools.aqua.stars.logic.kcmftbl.*
+import tools.aqua.stars.logic.kcmftbl.future.maxPrevalence
 
 /** This class tests the CMFTBL operator [maxPrevalence]. */
 class TestMaxPrevalence {